+++ AdvancedVision/NeuralNetworkTest/CsvParser.cpp
@@ -42,7 +42,7 @@ void CsvParser::bankNotesCsvtoTrainingSet(const std::string path, cv::Mat_<doubl
 		std::vector<std::string> vstrings(begin, end);
 
 		// Add input to inputmatrix, the first element in the row should be a 1
-		cv::Mat inputRow = (cv::Mat_<double>(1, vstrings.size() ) <<
+		cv::Mat inputRow = (cv::Mat_<double>(1, static_cast<int>(vstrings.size()) ) <<
 			1, std::stod(vstrings[VARIANCE_INDEX]), std::stod(vstrings[SKEWNESS_INDEX]), std::stod(vstrings[CURTOSIS_INDEX]), std::stod(vstrings[ENTROPY_INDEX]));
 		inputSet.push_back(inputRow);
